requirements.txt
# ============================================================
# Customer Churn Prediction — Dependencies
# ============================================================

# Core ML & Data
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.2.2
xgboost==2.0.3
imbalanced-learn==0.11.0
joblib==1.3.2

# Explainability
shap==0.44.0

# Visualisation
matplotlib==3.7.2
seaborn==0.13.1
plotly==5.18.0

# Dashboard
streamlit==1.32.0
pillow==10.0.1

# Notebooks
jupyter==1.0.0
nbformat==5.9.2
ipykernel==6.27.1

# Reporting
fpdf2==2.7.6
kaleido==0.2.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1